Venta controller, and some necesary changes with values Types

+++ src/main/java/com/explicacionD1/projectD1/Repository/DetalleVentaRepository.java
@@ -17,5 +17,5 @@ public interface DetalleVentaRepository extends JpaRepository<DetalleVenta, Long
 
     List<DetalleVenta> findBySubtotalLessThanEqual(Double monto);
 
-    List<DetalleVenta> findByCantidadLessThanEqual(Double cantidad);
+    List<DetalleVenta> findByCantidadLessThanEqual(Long cantidad);
 }
